APIGateway/requirements.txt
fastapi==0.89.1
pydantic==1.10.4
uvicorn==0.20.0
confluent-kafka==1.9.2
aiokafka==0.8.0
numpy==1.24.1
opencv-python==4.7.0.68
psycopg2-binary==2.9.5
sqlalchemy==1.4.46
python-multipart==0.0.5
pycryptodomex==3.16.0
python-dotenv==0.21.0
pytest==7.2.1
httpx==0.23.3

